Real-Time-Query-Workshop/python/code/requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
img2vec-pytorch==1.0.1
numpy==1.24.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.8.0.121
Pillow==9.4.0
redis==4.5.1
requests==2.28.2
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.14
